Move a `String` extension to separate file.
